+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Pashmak Changelog
 
+## 0.6, 0.6-alpha0 (2020-12-11)
+
+- removed `loop` and `while` functions
+- created **Structs** system
+- change default `RuntimeError` name to the real python exception name is exception system
+- created consts system
+
 ## 0.5.4 (2020-12-6)
 
 - fixed section system loop bug